fix: move to /utils
