+++ src/utils/page_anim_hotfix.js
